seek(2) returns and takes off_t, not int. Use the correct type.
This lets this interface work on systems with 64-bit off_t but 32-bit int
(which is most of them) when working with files larger than 2 GB.

+++ core/include/core/G3Reader.h
@@ -15,8 +15,8 @@ public:
                  float timeout = -1.);
 
 	void Process(G3FramePtr frame, std::deque<G3FramePtr> &out);
-	int Seek(int offset);
-	int Tell();
+	off_t Seek(off_t offset);
+	off_t Tell();
 
 private:
 	void StartFile(std::string path);
